[ts-scripts] add skip-image-deletion option to test-runner (#3816)
This PR makes the following changes:

- Add a new flag to ts-scripts test command called `skip-image-deletion`
which will skip any command to delete the docker cache or a docker image
within the docker cache.
- Add SKIP_IMAGE_DELETION variable to `scripts/src/helpers/config.ts`
-bump scripts from v1.4.2 to v1.5.0

ref: #3811

+++ packages/scripts/src/helpers/kind.ts
@@ -80,7 +80,7 @@ export class Kind {
 
     // TODO: check that image is loaded before we continue
     async loadServiceImage(
-        serviceName: string, serviceImage: string, version: string
+        serviceName: string, serviceImage: string, version: string, skipDelete: boolean
     ): Promise<void> {
         let subprocess;
         try {
@@ -95,7 +95,9 @@ export class Kind {
                 subprocess = await execaCommand(`gunzip -d ${filePath}`);
                 signale.info(`${subprocess.command}: successful`);
                 subprocess = await execaCommand(`kind load --name ${this.clusterName} image-archive ${tarPath}`);
-                fs.rmSync(tarPath);
+                if (!skipDelete) {
+                    fs.rmSync(tarPath);
+                }
             } else {
                 subprocess = await execaCommand(`kind load --name ${this.clusterName} docker-image ${serviceImage}:${version}`);
             }
